wisp/pyproject.toml
[project]
name = "wisp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-dotenv>=0.21.0",
    "mcp>=1.26.0",
    "requests>=2.32.5",
    "numpy>=2.4.1",
    "sqlite-vec>=0.1.6",
    "sentence-transformers>=5.2.2",
    "fastapi>=0.115.0",
    "uvicorn>=0.30.0",
    "fastmcp>=2.14.4",
    "rich>=14.3.1",
    "typer>=0.21.1",
    "redis>=7.1.0",
    "fakeredis>=2.33.0",
    "opentelemetry-api>=1.39.0",
    "opentelemetry-sdk>=1.39.0",
    "websockets>=16.0",
]
